Foundation Classes, TKMath - Rework numerical algorithms (#1479)
  - Introduce consistent zero-based APIs using size_t for dimensions and indices.
  - Add explicit result statuses, optional outputs, and callback error handling.
  - Improve scaling, validation, and convergence checks in numerical solvers.
  - Strengthen polynomial root isolation, multiplicity reporting, and deflation.
  - Improve integration, optimization, root-finding, and nonlinear-system methods.
  - Keep fixed-size Newton solvers allocation-free in their iteration loops.
  - Add zero-based accessors to math_Vector, math_Matrix, and NCollection_Array2.
  - Expand GTests and comparisons with equivalent legacy math_* algorithms.
  - Document indexing, failure behavior, tolerances, and convergence contracts.

+++ src/ModelingData/TKGeomBase/ExtremaPC/ExtremaPC_Ellipse.hxx
@@ -256,7 +256,7 @@ private:
     };
 
     // Add solutions from solver
-    for (int i = 0; i < aTrigRes.NbRoots; ++i)
+    for (size_t i = 0; i < aTrigRes.NbRoots; ++i)
     {
       addExtremum(aTrigRes.Roots[i]);
     }
